requirements.txt
Flask==2.3.0
Werkzeug==2.3
pymysql
cryptography
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Minify
Flask-QRcode

Flask-SQLAlchemy
Flask-WTF

html2text
pandas
python-dotenv
reportlab

git+https://github.com/ajrlewis/flask_wallet_of_satoshi.git
git+https://github.com/ajrlewis/bitcoinkit.git
